+++ svc/ChangelogService.ts
@@ -26,9 +26,8 @@ import {isEmpty, forOwn, includes} from 'lodash';
  *
  * Several additional options can be controlled via soft-config - see below.
  *
- * @see XH.showChangelog - public API for displaying the changelog, if enabled and populated.
- * @see whatsNewButton - utility button that conditionally renders when an unread entry exists for
- *      the currently deployed app version. Installed by default in desktop appBar.
+ * @see XH.showChangelog
+ * @see whatsNewButton
  */
 export class ChangelogService extends HoistService {
     override xhImpl = true;